AutoTrade/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
Django==4.0.3
django-environ==0.8.1
fonttools==4.32.0
idna==3.3
kiwisolver==1.4.2
matplotlib==3.5.1
mysqlclient==2.1.0
numpy==1.22.3
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
PyJWT==2.3.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
pyupbit==0.2.26
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
websockets==9.1
